02-aml-instance/deploy_requirements.txt
azure-ai-ml==1.32.0
azure-identity==1.25.3
ipykernel==7.2.0
pydantic==2.12.5
pydantic_settings==2.13.1
python-dotenv==1.2.2